[Silver I] Title: 진심 좌우 반복뛰기, Time: 36 ms, Memory: 2032 KB -BaekjoonHub

+++ 백준/Silver/23631. 진심 좌우 반복뛰기/진심 좌우 반복뛰기.cc
@@ -13,7 +13,7 @@ int main()
     cin >> T;
 
     for (int t = 0; t < T; t++) {
-        long long N, K;
+        int N, K;
         cin >> N >> K;
 
         // 찾고자 하는 mid 값은 층(?) 수
@@ -21,26 +21,11 @@ int main()
         // 2층 = 첫번째 왼쪽
         // 3층 = 두번째 오른쪽 
         // ...
-   //     long long start = 0;
-   //     long long end = N + 1;
-   //     while (start <= end) {
-   //         long long mid = (start + end) / 2;
-			//long long midSum = mid * (mid + 1);
-
-   //         if (midSum <= (N - 1) * 2 / K) { 
-   //             start = mid + 1;
-			//	//cout << " mid: " << mid << ", midSum: " << midSum << "\n";
-   //         }
-   //         else {
-   //             end = mid - 1;
-			//	//cout << " mid: " << mid << ", midSum: " << midSum << "\n";
-   //         }
-   //     }
 
         int find = floor((-1 + sqrt(1 + 8 * (N - 1) / K)) / 2);
 		//cout << "find: " << find << "\n";
 
-		long long remain = N - 1 - K * find * (find + 1) / 2;
+		int remain = N - 1 - K * find * (find + 1) / 2;
         //cout << "remain: " << remain << "\n";
 
         string dir;
@@ -51,7 +36,7 @@ int main()
             dir = "L";
         }
 
-        long long distance;
+        int distance;
         if (dir == "R") {
             distance = (find + 1) / -2 * K + remain;
         }
